feat(calendar): cải tiến hiển thị tiêu đề lịch - [style] Bỏ comment cho thuộc tính center trong headerToolbar feat(sections): cải tiến hiển thị thông tin trong bảng - [feat] Thêm render cho các cột trong SectionsColumns - [feat] Thêm render cho các cột trong SectionsSummaryColumns fix(calendar): sửa định dạng tháng trong tiêu đề lịch - [fix] Đổi month từ "long" sang "short" trong titleFormat

+++ src/component/config/calendar-config.js
@@ -12,7 +12,7 @@ export const CALENDAR_CONFIG = {
   eventBackgroundColor: "transparent",
   eventBorderColor: "transparent",
   firstDay: 7,
-  titleFormat: { year: "numeric", month: "long", day: "2-digit" },
+  titleFormat: { year: "numeric", month: "short", day: "2-digit" },
   eventTimeFormat: { hour: "2-digit", minute: "2-digit", hour12: false },
   buttonText: {
     today: "Hôm nay",
